gazebo_gym/requirements.txt
defusedxml==0.6.0
gym==0.17.3
mpi4py==3.0.3
nptyping==1.3.0
numpy-quaternion==2020.9.5.14.42.2
pkg-resources==0.0.0
pybullet==2.8.1
rospkg==1.2.8
tensorflow-gpu==2.3.0
tf-agents==0.6.0
tqdm==4.50.2
numpy==1.18.5


absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
catkin-pkg==0.4.23
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
decorator==4.4.2
distro==1.5.0
dm-tree==0.1.5
docutils==0.16
future==0.18.2
gast==0.3.3
gin-config==0.3.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
llvmlite==0.34.0
Markdown==3.3
numba==0.51.2
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-probability==0.11.1
termcolor==1.1.0
typish==1.8.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
